ai_bi_pipeline_project/requirements.txt
pandas
sqlalchemy
psycopg2-binary
snowflake-connector-python
scikit-learn
joblib
streamlit
apache-airflow==2.6.3